backend/requirements.txt
# Runtime dependencies for the KnowledgeDeck backend.
# Install in Docker via `pip install -r requirements.txt`.

alembic>=1.13.0
fastapi>=0.115.0
httpx>=0.27.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-text-splitters>=0.3.0
aiosqlite>=0.20.0
pydantic-settings>=2.4.0
pypdf>=5.0.0
python-docx>=1.1.0
python-pptx>=1.0.0
python-multipart>=0.0.20
qdrant-client>=1.12.0
sqlalchemy>=2.0.32
typer>=0.12.0
uvicorn[standard]>=0.30.0
